Okay... we did a lot here.
First... docker:
We made a config. in a way where
we have them both product-api and this
api running at the same time, we just
changed a few ports and the database
host needs to be our local ip address.

We have basic functionality (not fully tested)
for creating, deleting and consulting carts.

We have the addProducts functionality.
Seems that having all the migrations in
one place helps doing debugging and also
having a centralized place sounds like a
good idea.